move lifetime to a split file so that go1.1 is also work with this feature
